More longs

+++ elems/tadjgen.cpp
@@ -34,7 +34,7 @@ CTOR(TAdjGen,TMiso)
 }
 
 
-int TAdjGen::miso_startLoop( int /*acnx*/, int /*acny*/ )
+int TAdjGen::miso_startLoop( long /*acnx*/, long /*acny*/ )
 {
   ctt = ig = ig2 = dlt = tick_T = tick_old = av = av2 = av_diff = 0;
   currOut = 0; real_tick = 0; tick = 0;
