+++ elems/tadjgen.h
@@ -39,7 +39,7 @@ class TAdjGen : public TMiso  {
    /** main computation function */
    virtual double f() noexcept override;
    /** reimplemented from TMiso to reset */
-   virtual int miso_startLoop( int acnx, int acny );
+   virtual int miso_startLoop( long acnx, long acny );
 
    /** type of averaging, */
    PRM_LIST( type, efNRC, "Type", "Type of generator", "enum=GenType" );
