+++ elems/tcorralalysis.cpp
@@ -40,7 +40,7 @@ void TCorrAnalysis::reset_data()
   ave_x = ave_y = ave_x2 = ave_y2 = 0;
 }
 
-int TCorrAnalysis::miso_startLoop( int /*acnx*/, int /*acny*/ )
+int TCorrAnalysis::miso_startLoop( long /*acnx*/, long /*acny*/ )
 {
   reset_data();
   return 1;
